Split constants into core/ and http/
