+++ packages/core/src/sessions/http/remove-session-cookie.spec.ts
@@ -10,7 +10,7 @@ import {
   SESSION_DEFAULT_CSRF_COOKIE_NAME,
   SESSION_DEFAULT_SAME_SITE_ON_CSRF_ENABLED,
   SESSION_USER_COOKIE_NAME,
-} from '../constants';
+} from './constants';
 import { removeSessionCookie } from './remove-session-cookie';
 
 describe('removeSessionCookie', () => {
