+++ packages/core/src/sessions/http/remove-session-cookie.ts
@@ -7,7 +7,7 @@ import {
   SESSION_DEFAULT_CSRF_COOKIE_NAME,
   SESSION_DEFAULT_SAME_SITE_ON_CSRF_ENABLED,
   SESSION_USER_COOKIE_NAME,
-} from '../constants';
+} from './constants';
 
 /**
  * Deletes the browser cookie containing the session token.
