+++ packages/core/src/sessions/http/set-session-cookie.spec.ts
@@ -10,7 +10,7 @@ import {
   SESSION_DEFAULT_CSRF_COOKIE_NAME,
   SESSION_DEFAULT_SAME_SITE_ON_CSRF_ENABLED,
   SESSION_USER_COOKIE_NAME,
-} from '../constants';
+} from './constants';
 import { Session } from '../core/session';
 import { setSessionCookie } from './set-session-cookie';
