+++ packages/core/src/sessions/http/use-sessions.hook.spec.ts
@@ -24,11 +24,13 @@ import {
   ServiceManager
 } from '../../core';
 import {
-  SESSION_DEFAULT_ABSOLUTE_TIMEOUT,
   SESSION_DEFAULT_COOKIE_NAME,
-  SESSION_DEFAULT_INACTIVITY_TIMEOUT,
   SESSION_USER_COOKIE_NAME
-} from '../constants';
+} from './constants';
+import {
+  SESSION_DEFAULT_ABSOLUTE_TIMEOUT,
+  SESSION_DEFAULT_INACTIVITY_TIMEOUT,
+} from '../core/constants'
 import { readSession } from '../core/read-session';
 import { Session } from '../core/session';
 import { SessionState } from '../core/session-state.interface';
